treat each gene separately, and exclude variants not in known genes
Previously the code would check each of the genes that a CNV overlaps. This is now
used for SNVs as well, which will occur when we have variants that have been picked
up by checking overlap to the known genes.

In order to exclude mismatches, if using the known genes, for each variant we check
if the gene is within the known genes. This should exclude some of the variants
that overlap known genes, but which do not have the appropriate inheritance type for
that known gene.