fixed a bug when excluding compound hets based on polyphen predictions
We exclude compound hets when one or more of the pair is predicted as
benign by polyphen. This needed to be updated for the shift to per-gene
consequences/predictions, as I found a case where a variant was excluded
it had polyphen prediction of benign, but not for the gene that it was
filtered on. This had left the partner variant in the filter output,
since I had accounted for gene-specific predictions when looking at the
partner variants.

This has now been updated to only look at the predictions for the
gene/s of interest.