abstracted out handling gene symbols
We need to match variants to known genes by HGNC_IDs, but to be able to
use
the corresponding HGNC symbols in VCF and text outputs. I've set up a
class to
pass in the gene symbols from the INFO for a given allele. This can pull
out
symbols based on our priority of symbol types.
Updated various unit tests to match new behaviour.