fixed bug when looking for symbols matches with '*' alleles
There was a bug when logging the variants for a gene. For logging
purposes I prefer HGNC symbols to HGNC IDs, so I preferentially convert
the HGNC ID to a HGNC symbol from the symbols for the first alt allele.
However, in multi-allelic sites where the first allele is ‘*’, there are
no annotated gene symbols to match up against, thus raising the
KeyError.

This now checks symbols for the other alleles when this error occurs.